requirements.txt
APScheduler==3.10.4
flask==2.3.3
Flask-Cors==4.0.0
libsass==0.22.0
NESP-Lib==1.0.2
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
git+https://github.com/fdobrovolny/RS485-relay-board.git
